account_manager/pyproject.toml
[tool.poetry]
name = "crypto-acc-manager"
version = "0.1.0"
description = ""
authors = ["usik-sviatoslav <usik.sviatoslav.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pydantic = "^2.5.2"
django = "^5.0.6"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
django-debug-toolbar = "^4.3.0"
python-multipart = "^0.0.9"
cryptography = "^42.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
